Language.Fixpoint.Partition

Description

This module implements functions that print out statistics about the constraints.

Documentation

data CPart a

Constraint Partition Container -------------------------------------

Constructors

CPart

Fields

pws :: HashMap KVar (WfC a)

pcm :: HashMap Integer (SubC a)

cFileName :: FilePath

Split constraints

partition :: Fixpoint a => Config -> FInfo a -> IO (Result a)

partition'

Arguments

:: Maybe MCInfo
Nothing to produce the maximum possible number of partitions. Or a MultiCore Info to control the partitioning

-> FInfo a

-> (KVGraph, [FInfo a])

Partition an FInfo into multiple disjoint FInfos

partitionN

Arguments

:: MCInfo
describes thresholds and partiton amounts

-> FInfo a
The originial FInfo

-> [CPart a]
A list of the smallest possible CParts

-> [FInfo a]
At most N partitions of at least thresh work

Partition an FInfo into a specific number of partitions of roughly equal amounts of work

Information about cores

data MCInfo

Multicore info -----------------------------------------------------

Constructors

MCInfo

Fields

mcCores :: Int

mcMinPartSize :: Int

mcMaxPartSize :: Int

mcInfo :: Config -> IO MCInfo