Language.Fixpoint.Smt.Serialize

Description

This module contains the code for serializing Haskell values into SMTLIB2 format, that is, the instances for the SMTLIB2 typeclass. We split it into a separate module as it depends on Theories (see smt2App).

Documentation

smt2Bop :: Bop -> Expr -> Expr -> Text

uOp :: Bop -> Located Symbol

smt2App :: LocSymbol -> [Expr] -> Text

smt2App' :: SMTLIB2 a => a -> [Text] -> Text

mkRel :: (SMTLIB2 a, SMTLIB2 a1) => Brel -> a -> a1 -> Text

mkNe :: (SMTLIB2 a, SMTLIB2 a1) => a -> a1 -> Text

smt2s :: SMTLIB2 a => [a] -> Text

smt2many :: [Text] -> Text