Language.Fixpoint.Solver

Description

This module implements the top-level API for interfacing with Fixpoint In particular it exports the functions that solve constraints supplied either as .fq files or as FInfo.

Invoke Solver on an FInfo

solve :: (NFData a, Fixpoint a) => Solver a

Solve FInfo system of horn-clause constraints ------------------------

Invoke Solver on a .fq file

solveFQ :: Config -> IO ExitCode

Solve an .fq file ----------------------------------------------------

Function to determine outcome

resultExit :: FixResult a -> ExitCode

Extract ExitCode from Solver Result ----------------------------------

Parse Qualifiers from File

parseFInfo :: [FilePath] -> IO (FInfo a)

Parse External Qualifiers --------------------------------------------