Language.Fixpoint.Types.Refinements

Description

This module contains the data types for representing terms in the refinement logic; currently split into Expr and Pred but which will be unified.

Representing Terms

data SymConst

Expressions ---------------------------------------------------------------

Uninterpreted constants that are embedded as "constant symbol : Str"

Constructors

SL !Text

data Constant

Constructors

I !Integer

R !Double

L !Text !Sort

data Bop

Constructors

Plus

Minus

Times

Div

Mod

data Brel

Constructors

Eq

Ne

Gt

Ge

Lt

Le

Ueq

Une

data Expr

Constructors

ESym !SymConst

ECon !Constant

EVar !Symbol

EApp !LocSymbol ![Expr]

ENeg !Expr

EBin !Bop !Expr !Expr

EIte !Expr !Expr !Expr

ECst !Expr !Sort

EBot

ETApp !Expr !Sort

ETAbs !Expr !Symbol

PTrue

PFalse

PAnd !(ListNE Expr)

POr ![Expr]

PNot !Expr

PImp !Expr !Expr

PIff !Expr !Expr

PAtom !Brel !Expr !Expr

PKVar !KVar !Subst

PAll ![(Symbol, Sort)] !Expr

PExist ![(Symbol, Sort)] !Expr

PTop

newtype KVar

Kvars ---------------------------------------------------------------------

Constructors

KV

Fields

kv :: Symbol

newtype Subst

Substitutions -------------------------------------------------------------

Constructors

Su (HashMap Symbol Expr)

newtype Reft

Constructors

Reft (Symbol, Expr)

data SortedReft

Constructors

RR

Fields

sr_sort :: !Sort

sr_reft :: !Reft

Constructing Terms

eVar :: Symbolic a => a -> Expr

elit :: Located Symbol -> Sort -> Expr

eProp :: Symbolic a => a -> Expr

pAnd :: ListNE Expr -> Expr

pOr :: ListNE Expr -> Expr

pIte :: Expr -> Expr -> Expr -> Expr

isTautoPred :: Expr -> Bool

Generalizing Embedding with Typeclasses

class Expression a where

Generalizing Symbol, Expression, Predicate into Classes -----------

Values that can be viewed as Constants

Values that can be viewed as Expressions

Methods

expr :: a -> Expr

class Predicate a where

Values that can be viewed as Predicates

Methods

prop :: a -> Expr

class Subable a where

Class Predicates for Valid Refinements -----------------------------

Minimal complete definition

syms, substa, substf, subst

Methods

syms :: a -> [Symbol]

substa :: (Symbol -> Symbol) -> a -> a

substf :: (Symbol -> Expr) -> a -> a

subst :: Subst -> a -> a

subst1 :: a -> (Symbol, Expr) -> a

class (Monoid r, Subable r) => Reftable r where

Minimal complete definition

isTauto, ppTy, bot, toReft, ofReft, params

Methods

isTauto :: r -> Bool

ppTy :: r -> Doc -> Doc

top :: r -> r

bot :: r -> r

meet :: r -> r -> r

toReft :: r -> Reft

ofReft :: Reft -> r

params

Arguments

:: r

-> [Symbol]
parameters for Reft, vv + others

Constructing Refinements

reft :: Symbol -> Expr -> Reft

trueSortedReft :: Sort -> SortedReft

trueReft :: Reft

falseReft :: Reft

exprReft :: Expression a => a -> Reft

notExprReft :: Expression a => a -> Reft

uexprReft :: Expression a => a -> Reft

symbolReft :: Symbolic a => a -> Reft

Generally Useful Refinements --------------------------

usymbolReft :: Symbolic a => a -> Reft

propReft :: Predicate a => a -> Reft

predReft :: Predicate a => a -> Reft

reftPred :: Reft -> Expr

reftBind :: Reft -> Symbol

isFunctionSortedReft :: SortedReft -> Bool

Refinements ----------------------------------------------

functionSort :: Sort -> Maybe (Int, [Sort], Sort)

isNonTrivial :: Reftable r => r -> Bool

isSingletonReft :: Reft -> Maybe Expr

Predicates ----------------------------------------------------------------

isEVar :: Expr -> Bool

isFalse :: Falseable a => a -> Bool

flattenRefas :: [Expr] -> [Expr]

conjuncts :: Expr -> [Expr]

mapPredReft :: (Expr -> Expr) -> Reft -> Reft

pprintReft :: Reft -> Doc

reftConjuncts :: Reft -> [Reft]

intKvar :: Integer -> KVar

vv_ :: Symbol