Language.Fixpoint.Types.Sorts

Description

This module contains the data types, operations and serialization functions for representing Fixpoint's implication (i.e. subtyping) and well-formedness constraints in Haskell. The actual constraint solving is done by the `fixpoint.native` which is written in Ocaml.

Embedding to Fixpoint Types

data Sort

Constructors

FInt

FReal

FNum
numeric kind for Num tyvars

FFrac
numeric kind for Fractional tyvars

FObj Symbol
uninterpreted type

FVar !Int
fixpoint type variable

FFunc !Int ![Sort]
type-var arity, in-ts ++ [out-t]

FTC FTycon

FApp Sort Sort
constructed type

newtype Sub

Constructors

Sub [(Int, Sort)]

data FTycon

type TCEmb a = HashMap a FTycon

sortFTycon :: Sort -> Maybe FTycon

intFTyCon :: FTycon

boolFTyCon :: FTycon

realFTyCon :: FTycon

numFTyCon :: FTycon

intSort :: Sort

Exported Basic Sorts -----------------------------------------------

realSort :: Sort

Exported Basic Sorts -----------------------------------------------

boolSort :: Sort

Exported Basic Sorts -----------------------------------------------

strSort :: Sort

Exported Basic Sorts -----------------------------------------------

funcSort :: Sort

Exported Basic Sorts -----------------------------------------------

listFTyCon :: FTycon

isListTC :: FTycon -> Bool

fTyconSymbol :: FTycon -> Located Symbol

symbolFTycon :: LocSymbol -> FTycon

fTyconSort :: FTycon -> Sort

fApp :: Sort -> [Sort] -> Sort

fApp' :: Sort -> ListNE Sort

fAppTC :: FTycon -> [Sort] -> Sort

fObj :: LocSymbol -> Sort

sortSubst :: HashMap Symbol Sort -> Sort -> Sort

functionSort :: Sort -> Maybe (Int, [Sort], Sort)