Language.Fixpoint.Types.Substitutions

Description

This module contains the various instances for Subable, which (should) depend on the visitors, and hence cannot be in the same place as the Term definitions.

Documentation

mkSubst :: [(Symbol, Expr)] -> Subst

isEmptySubst :: Subst -> Bool

substExcept :: Subst -> [Symbol] -> Subst

substfExcept :: (Symbol -> Expr) -> [Symbol] -> Symbol -> Expr

subst1Except :: Subable a => [Symbol] -> a -> (Symbol, Expr) -> a

targetSubstSyms :: Subst -> [Symbol]